blacklist.conf: blacklisted f4cb767d76cf7ee72f97dd76f6cfa6c76a5edc89
This patch is non-trivial to backport. Moreover this is VM_BUG_ON which
is triggered and that is NOOP on our trees. So even if we have somebody
nasty trying to fool us then it would simply shoot its own feet and
mmap would fail later in the call chain if conflicting with an existing
one.

suse-commit: e8b8a14b8b51b74f48344ae6c7c40496f7cd357f