btrfs: fix backport error in submit_stripe_bio (bsc#1114763).
There was an error in
patches.suse/btrfs-0517-convert-pr_-to-btrfs_-where-possible.patch where
we weren't passing the fs_info to btrfs_debug.  Let's clean that up to
match the upstream commit.

There is no end-user impact.  This code is protected by #ifdef DEBUG and
is not built on shipped kernels.  It would've been noisy at build time
otherwise.

suse-commit: 890be49bf438908bcf171b4b893153cf6adcc65c